Calculates the final energy to useful exergy multipliers

Calculates the final energy to useful exergy multipliers

Usage

calc_Ef_to_Xu_exiobase(
  EtafuYEIOU_mats,
  phi_vecs,
  eta_fu_phi_Y_EIOU_agg,
  years_exiobase,
  full_list_exiobase_flows,
  country_concordance_table_df,
  useful_energy_flow = "Useful.energy.flow",
  exiobase_flow = "Exiobase.Flow",
  pfu_code = "PFU.code",
  pfu_flow = "PFU.flow",
  iea_country_name = "IEA.country.name",
  iea_country_name_accented = "IEA.name.accented",
  phi = "phi",
  matnames = "matnames",
  colnames = "colnames",
  matvals = "matvals",
  rowtypes = "rowtypes",
  coltypes = "coltypes",
  energy_type_E = IEATools::energy_types$e,
  country = IEATools::iea_cols$country,
  product = IEATools::iea_cols$product,
  year = IEATools::iea_cols$year,
  flow = IEATools::iea_cols$flow,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  eta_p_eiou = "eta_p_eiou",
  eta_p_eiou_y = "eta_p_eiou_y",
  eta_fu_Y_E = "eta_fu_Y_E",
  eta_fu_EIOU_E = "eta_fu_EIOU_E",
  eta_fu_Y_X = "eta_fu_Y_X",
  eta_fu_EIOU_X = "eta_fu_EIOU_X",
  eta_phi_p_eiou = "eta_phi_p_eiou",
  eta_phi_p_eiou_y = "eta_phi_p_eiou_y",
  phi_eta_X = "phi_eta_X",
  eta = "eta"
)

Arguments

EtafuYEIOU_mats

The input data frame containing matrices with all the efficiencies by final demand sector and energy industry

phi_vecs

A data frame of phi (exergy-to-energy ratio) coefficients.

eta_fu_phi_Y_EIOU_agg

A data frame containing matrices with the economy-wide efficiencies time phi values by energy product

years_exiobase

The years for which the coefficients are provided to the Exiobase team

full_list_exiobase_flows

The full list of energy flows used in the Exiobase pipeline led by KR

country_concordance_table_df

A data frame containing the country concordance table

useful_energy_flow

The name of the column stating whether a flow is a final energy flow or not

exiobase_flow

The name of the column stating the name of the Exiobase flow

pfu_code

The name of the column containing the PFU country name

pfu_flow

The name of the column containing the PFU flow names

iea_country_name

The name of the column containing the IEA country name

iea_country_name_accented

The name of the column containing the IEA country name with accents

phi

The name of the column containing the phi values

matnames

The name of the column containing matrices names after unpacking the matrices

colnames

The name of the column containing the column names after unpacking the matrices

matvals

The name of the column containing matrices values after unpacking the matrices

rowtypes

The name of the column containing the matrices row types names after unpacking the matrices

coltypes

The name of the column containing matrices column types after unpacking the matrices

energy_type_E

The letter standing for energy as energy type.

country, year, product, flow, method, energy_type, last_stage

See IEATools::iea_cols.

eta_p_eiou

The name of the column containing the efficiencies.

eta_p_eiou_y

The name of the column containing the efficiencies of products when used as part of the final demand

eta_fu_Y_E

The name of the the column containing the efficiencies of products when used as part of final demand.

eta_fu_EIOU_E

The name of the column containing the efficiencies of products when used as part of the EIOU matrix.

eta_fu_Y_X

The name of the column containing the exergy efficiencies for products used as part of final demand.

eta_fu_EIOU_X

The name of the column containing the exergy efficiencies for products used as part of EIOU.

eta_phi_p_eiou

The name of the column containing the efficiency values multiplied by the phi values for products used in EIOU.

eta_phi_p_eiou_y

The name of the column containing the efficiency values multiplied by the phi values for products used economy-wide.

phi_eta_X

The name of the column containing the efficiencies, multiplied by the phi values, of products when used as part of the EIOU matrix.

eta

The name of the column containing the efficiencies.

Value

A data frame of the final energy to useful exergy multipliers