Calculates the aggregated FU efficiency of products when used in EIOU, Y, or economy-wide

Calculates the aggregated FU efficiency of products when used in EIOU, Y, or economy-wide

Usage

calc_fu_Y_EIOU_agg_efficiencies(
  C_mats_agg,
  eta_fu_vecs,
  phi_vecs,
  non_energy_use_machine = "Non-energy consumption -> NEU",
  eta.fu = "eta.fu",
  C_EIOU_agg = "C_EIOU_agg",
  C_Y_agg = "C_Y_agg",
  C_EIOU_Y_agg = "C_EIOU_Y_agg",
  C_EIOU_agg_excl_NEU = "C_EIOU_agg_excl_NEU",
  C_Y_agg_excl_NEU = "C_Y_agg_excl_NEU",
  C_EIOU_Y_agg_excl_NEU = "C_EIOU_Y_agg_excl_NEU",
  eta_p_eiou = "eta_p_eiou",
  eta_p_y = "eta_p_y",
  eta_p_eiou_y = "eta_p_eiou_y",
  phi = "phi",
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product
)

Arguments

C_mats_agg

A data frame containing the aggregated C matrices.

eta_fu_vecs

A data frame containing the efficiency vectors.

phi_vecs

A data frame containing the phi values.

non_energy_use_machine

The character string of the non-energy use machine that needs to be excluded for efficiencies excluding non-energy uses.

eta.fu

The name of the column containing the machine efficiencies in the eta_fu_vecs data frame.

C_EIOU_agg

The name of the column containing the C_EIOU aggregated matrix in the output data frame.

C_Y_agg

The name of the column containing the C_Y aggregated matrix in the output data frame.

C_EIOU_Y_agg

The name of the column containing the C_EIOU_Y aggregated matrix in the output data frame.

C_EIOU_agg_excl_NEU

The name of a temporary column containing the C_EIOU aggregated matrix excluding non-energy uses.

C_Y_agg_excl_NEU

The name of a temporary column containing the C_Y aggregated matrix excluding non-energy uses.

C_EIOU_Y_agg_excl_NEU

The name of a temporary column containing the C_EIOU_Y aggregated matrix excluding non-energy uses.

eta_p_eiou

The name of the column containing the efficiency of each product when used in EIOU.

eta_p_y

The name of the column containing the efficiency of each product when used in final demand.

eta_p_eiou_y

The name of the column containing the efficiency of each product when used in either EIOU or final demand.

phi

The name of the column containing the phi values.

country, method, energy_type, last_stage, year, product

See IEATools::iea_cols.

Value

A data frame containing the product efficiencies when used in EIOU, Y, or economy-wide (Y and EIOU).