Create an analysis pipeline

The pipeline a targets pipeline that creates the PFU database.

Usage

get_pipeline(
  countries = "all",
  additional_exemplar_countries = NULL,
  specify_non_energy_flows = FALSE,
  matrix_class = c("matrix", "Matrix"),
  apply_fixes = FALSE,
  years = "all",
  how_far = "all_targets",
  iea_data_path,
  country_concordance_path,
  mw_concordance_path,
  amw_analysis_data_path,
  hmw_analysis_data_path,
  phi_constants_path,
  ceda_data_folder,
  fao_data_path,
  ilo_employment_data_path,
  ilo_working_hours_data_path,
  machine_data_path,
  exemplar_table_path,
  fu_analysis_folder,
  exiobase_energy_flows_path,
  years_exiobase,
  reports_source_folders,
  reports_dest_folder,
  pipeline_releases_folder,
  pipeline_caches_folder,
  release = FALSE
)

Arguments

countries

A vector of abbreviations for countries whose energy conversion chain is to be analyzed, such as "c('GHA', 'ZAF')". Countries named in countries can also serve as exemplars for final-to-useful allocations and efficiencies.

additional_exemplar_countries

A vector of country abbreviations for which final-to-useful allocations and efficiencies will be read. An energy conversion chain will not be constructed for these countries. However, their final-to-useful allocations and efficiencies may be used as exemplar information for the countries in countries. Default is NULL, indicating no additional exemplars.

specify_non_energy_flows

A boolean that tells whether to provide additional specificity to non-energy flows, when available. Default is FALSE.

matrix_class

One of "matrix" or "Matrix", specifying which kind of matrix objects are to be created. Default is "matrix".

apply_fixes

A boolean that tells whether to fix some of the IEA WEEB data, where possible. Default is FALSE.

years

The years to be studied.

how_far

A string indicating the last target to include in the plan that is returned. Default is "all_targets" to indicate all targets of the plan should be returned.

iea_data_path

The path to IEA extended energy balance data in .csv format.

country_concordance_path

The path to the country concordance Excel file.

mw_concordance_path

The path to the muscle work concordance file.

amw_analysis_data_path

The path to the animal muscle work data file.

hmw_analysis_data_path

The path to the human muscle work data file.

phi_constants_path

The path to a phi (exergy-to-energy ratio) Excel file.

ceda_data_folder

The path to the CEDA data in text file, .per, format.

fao_data_path

The path to Food and Agriculture Organization (FAO) data.

ilo_employment_data_path

The path to International Labor Organization (ILO) employment data.

ilo_working_hours_data_path

The path to International Labor Organization (ILO) working hours data.

machine_data_path

The path to the machine data in .xlsx format.

exemplar_table_path

The path to an exemplar table.

fu_analysis_folder

The path to a folder containing final-to-useful analyses. Sub-folders named with 3-letter country abbreviations are assumed.

exiobase_energy_flows_path

The path to the file where the list of Exiobase energy flows, and their concordance to the PFU database flows are stored.

years_exiobase

The years for which the multipliers to provide to the Exiobase team need to be calculated.

reports_source_folders

A string vector containing paths to folders of report sources, usually .Rnw or .Rmd files.

reports_dest_folder

The path to a folder into which reports are written.

pipeline_releases_folder

The path to a folder where releases of important targets are stored for later retrieval as pinned items on a pinboard.

pipeline_caches_folder

The path to a folder where .zipped versions of the pipeline cache are stored.

release

A boolean that tells whether a new release of the PSUT targets should be made. Default is FALSE.

Value

A targets pipeline.